Get my level to work with Typescript Example
This commit adds all the files from the phaserTest example project
To this Typescript example project. With these changes running:
npm run watch

Should cause the game to be compiled and run on http://localhost:10001
and continue to be updated as the code is changed.
The starting scene of the game is the StartLevel Scene